@@ -7,7 +7,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -15,24 +15,29 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   
-  dio: ^4.0.6
-  provider: ^6.0.2
+  dio: ^4.0.6  
+  lottie: ^1.3.0
+  get_it: ^7.2.0
   equatable: ^2.0.3
-  connectivity_plus: ^2.3.0
-  shared_preferences: ^2.0.13
+  connectivity_plus: ^2.3.5
+  shared_preferences: ^2.0.15
 
-  firebase_core: ^1.15.0
-  firebase_analytics: ^9.1.6
-  firebase_crashlytics: ^2.6.3
-  firebase_remote_config: ^2.0.5
+  bloc: ^8.0.3
+  flutter_bloc: ^8.0.1
+  bloc_concurrency: ^0.2.0
+
+  firebase_core: ^1.19.1
+  firebase_analytics: ^9.1.12
+  firebase_crashlytics: ^2.8.4
+  firebase_remote_config: ^2.0.11
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   mocktail: ^0.3.0
-  
-  flutter_lints: ^1.0.0
+  bloc_test: ^9.0.3
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
@@ -43,3 +48,5 @@ flutter:
   assets:
     # Add assets from the images directory to the application.
     - assets/images/
+    # Add assets from the animations
+    - assets/animations/
