@@ -1,27 +1,28 @@
 import 'package:seller/src/modules/auth/data/models/me/me_model.dart';
-
 import 'package:seller/src/modules/http/data/clients/http_client.dart';
-import 'package:seller/src/modules/http/domain/helpers/http_helper.dart';
+
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
+
+import 'package:seller/src/core/domain/entities/exception/exception.dart';
+import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 
 import 'package:seller/src/modules/auth/domain/entities/me/me_entity.dart';
 import 'package:seller/src/modules/auth/domain/usecases/me/me_usecase.dart';
 
-class RemoteMeUseCase extends MeUseCase {
-  late final HttpClient _call;
+class RemoteMeUseCase implements MeUseCase {
+  late final HttpClient _httpClient;
 
-  RemoteMeUseCase(this._call);
+  RemoteMeUseCase(this._httpClient);
 
   @override
-  Future<MeEntity> getMe(MeParams params) async {
-    try {
-      final _response = await _call(
-        url: params.url,
-        method: HttpMethod.get,
-      );
+  Future<Either<HttpException, MeEntity>> call(
+    HttpParams params,
+  ) async {
+    final response = await _httpClient.get(params: params);
 
-      return MeModel.fromJson(_response.right);
-    } on HttpResponse catch (_) {
-      rethrow;
-    }
+    return response.either(
+      (left) => left,
+      (right) => MeModel.fromJson(right),
+    );
   }
 }
