@@ -1,157 +1,61 @@
-import 'dart:convert';
-
 import 'package:dio/dio.dart';
 
-import 'package:seller/src/utils/constants/app_constant.dart';
-
-import 'package:seller/src/core/domain/entities/exception/exception.dart';
-
 import 'package:seller/src/core/domain/entities/either/left_entity.dart';
+import 'package:seller/src/core/domain/entities/exception/exception.dart';
 import 'package:seller/src/core/domain/entities/either/right_entity.dart';
 import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 
-import 'package:seller/src/modules/storage/domain/usecases/storage_usecase.dart';
-
 import 'package:seller/src/modules/http/data/clients/http_client.dart';
-import 'package:seller/src/modules/http/domain/helpers/http_helper.dart';
-import 'package:seller/src/modules/http/external/interceptors/auth_interceptor.dart';
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
+import 'package:seller/src/modules/http/data/clients/config_dio_client.dart';
 
 class HttpAdapter implements HttpClient {
   late final Dio _http;
-  late final StorageUsecase _storageUseCase;
+  late final ConfigDioClient _dioAdapter;
 
   HttpAdapter(
     this._http,
-    this._storageUseCase,
+    this._dioAdapter,
   ) {
-    registerInterceptors();
+    configureHttp();
   }
 
   @override
-  void registerInterceptors() {
-    _http.interceptors.add(AuthInterceptor(_storageUseCase));
+  void configureHttp() {
+    _http.options = _dioAdapter.buildDioSettings();
+    _http.interceptors.addAll(_dioAdapter.builInterceptors());
   }
 
   @override
-  Future<Either<HttpException, Map<String, dynamic>>> call({
-    required String url,
-    required HttpMethod method,
-    Map<String, dynamic>? body,
-    Map<String, String>? headers,
+  Future<Either<HttpException, Map<String, dynamic>>> get({
+    required HttpParams params,
   }) async {
-    Future<Response>? _futureResponse;
-    Response<dynamic> _response = Response(
-      requestOptions: RequestOptions(
-        path: url,
-        sendTimeout: 6,
-        headers: _buildHeaders(headers),
-        responseType: ResponseType.json,
-      ),
-    );
-
-    final _jsonBody = _buildBody(body);
-    final _defaultHeaders = _buildHeaders(headers);
-
     try {
-      switch (method) {
-        case HttpMethod.post:
-          _futureResponse = _http.post(
-            url,
-            data: _jsonBody,
-            options: Options(headers: _defaultHeaders),
-          );
-          break;
-        case HttpMethod.get:
-          _futureResponse = _http.get(
-            url,
-            options: Options(headers: _defaultHeaders),
-          );
-          break;
-        case HttpMethod.put:
-          _futureResponse = _http.put(
-            url,
-            data: _jsonBody,
-            options: Options(headers: _defaultHeaders),
-          );
-          break;
-        case HttpMethod.patch:
-          _futureResponse = _http.patch(
-            url,
-            data: _jsonBody,
-            options: Options(headers: _defaultHeaders),
-          );
-          break;
-        case HttpMethod.delete:
-          _futureResponse = _http.delete(
-            url,
-            options: Options(headers: _defaultHeaders),
-          );
-          break;
-      }
+      final response = await _http.get(
+        params.url,
+        options: _dioAdapter.buildOptions(params),
+      );
 
-      _response = await _futureResponse;
+      return Right(_dioAdapter.buildResponseSuccess(response));
     } on DioError catch (error) {
-      return Left(_buildResponseError(error));
-    }
-
-    return Right(_buildResponseSuccess(_response));
-  }
-
-  Map<String, String> _buildEmptyBody() => {
-        'response': 'ok',
-      };
-
-  String? _buildBody(Map<String, dynamic>? body) =>
-      body != null ? jsonEncode(body) : null;
-
-  Map<String, String> _buildHeaders(Map<String, String>? headers) =>
-      headers?.cast<String, String>() ?? {}
-        ..addAll({
-          'accept': 'application/json',
-          'content-type': 'application/json',
-          'apikey': AppContants.apikeyAnon,
-        });
-
-  Map<String, dynamic> _buildResponseSuccess(
-    Response<dynamic> response,
-  ) {
-    switch (response.statusCode) {
-      case 200:
-        return response.data.isEmpty ? _buildEmptyBody() : response.data;
-      case 204:
-        return _buildEmptyBody();
-      default:
-        return _buildEmptyBody();
+      return Left(_dioAdapter.buildResponseError(error));
     }
   }
 
-  HttpException _buildResponseError(DioError error) {
-    switch (error.response?.statusCode ?? 500) {
-      case 400:
-        return HttpException(
-          code: HttpResponse.badRequest,
-          message: HttpResponse.badRequest.value,
-        );
-      case 401:
-        return HttpException(
-          code: HttpResponse.unauthorized,
-          message: HttpResponse.unauthorized.value,
-        );
-      case 403:
-        return HttpException(
-          code: HttpResponse.forbidden,
-          message: HttpResponse.forbidden.value,
-        );
-      case 404:
-        return HttpException(
-          code: HttpResponse.notFound,
-          message: HttpResponse.notFound.value,
-        );
-      default:
-        return HttpException(
-          code: HttpResponse.serverError,
-          message: HttpResponse.serverError.value,
-        );
+  @override
+  Future<Either<HttpException, Map<String, dynamic>>> post({
+    required HttpParams params,
+  }) async {
+    try {
+      final response = await _http.post(
+        params.url,
+        data: _dioAdapter.buildBody(params.body),
+        options: _dioAdapter.buildOptions(params),
+      );
+
+      return Right(_dioAdapter.buildResponseSuccess(response));
+    } on DioError catch (error) {
+      return Left(_dioAdapter.buildResponseError(error));
     }
   }
 }
