@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:seller/src/utils/localization/localization_config.dart';
 
 import 'package:seller/src/core/presentation/routes/routes.dart';
-import 'package:seller/src/core/presentation/widgets/provider_widget.dart';
 import 'package:seller/src/core/presentation/extensions/context_extension.dart';
 
 import 'package:seller/src/modules/splash/presentation/screens/splash_screen.dart';
@@ -13,21 +12,19 @@ class App extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ProviderWidget(
-      child: MaterialApp(
-        restorationScopeId: 'app',
-        theme: ThemeData(
-          useMaterial3: true,
-        ),
-        themeMode: ThemeMode.system,
-        debugShowCheckedModeBanner: false,
-        supportedLocales: supportedLocales,
-        localizationsDelegates: localizationsDelegates,
-        initialRoute: SplashScreen.path,
-        navigatorKey: Navigation.navigatorKey,
-        onGenerateRoute: Navigation.onGenerateRoute,
-        onGenerateTitle: (BuildContext context) => context.localize.appTitle,
+    return MaterialApp(
+      restorationScopeId: 'app',
+      theme: ThemeData(
+        useMaterial3: true,
+        primarySwatch: Colors.red,
       ),
+      themeMode: ThemeMode.system,
+      debugShowCheckedModeBanner: false,
+      supportedLocales: supportedLocales,
+      localizationsDelegates: localizationsDelegates,
+      initialRoute: SplashScreen.pathRoot,
+      onGenerateRoute: Navigation.onGenerateRoute,
+      onGenerateTitle: (BuildContext context) => context.localize.appTitle,
     );
   }
 }
