@@ -1,3 +1,7 @@
+import 'package:get_it/get_it.dart';
+
+import 'package:seller/src/core/external/di/module.dart';
+
 import 'package:seller/src/modules/auth/external/di/injection.dart';
 import 'package:seller/src/modules/http/external/di/injection.dart';
 import 'package:seller/src/modules/splash/external/di/injection.dart';
@@ -5,11 +9,18 @@ import 'package:seller/src/modules/storage/external/di/injection.dart';
 import 'package:seller/src/modules/firabase/external/di/injection.dart';
 import 'package:seller/src/modules/connectivity/external/di/injection.dart';
 
-final providers = [
-  ...storageModule,
-  ...httpModule,
-  ...connectivityModule,
-  ...firebaseModule,
-  ...splashModule,
-  ...authModule,
-];
+class ServiceLocator implements Module {
+  final GetIt locator;
+
+  ServiceLocator(this.locator);
+
+  @override
+  Future<void> initialized() async {
+    FirebaseModule(locator).initialized();
+    StorageModule(locator).initialized();
+    HttpModule(locator).initialized();
+    ConnectivityModule(locator).initialized();
+    SplashModule(locator).initialized();
+    AuthModule(locator).initialized();
+  }
+}
