@@ -1,11 +1,13 @@
 import 'package:flutter/cupertino.dart';
 
 mixin StateMixin<T extends StatefulWidget> on State<T> {
-  void completeState();
+  void onCreated();
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance?.addPostFrameCallback((_) => completeState());
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      onCreated();
+    });
   }
 }
