@@ -1,13 +1,9 @@
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
+
+import 'package:seller/src/core/domain/entities/exception/exception.dart';
+import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 import 'package:seller/src/modules/auth/domain/entities/me/me_entity.dart';
 
 abstract class MeUseCase {
-  Future<MeEntity> getMe(MeParams params);
-}
-
-class MeParams {
-  final String url;
-
-  const MeParams({
-    required this.url,
-  });
+  Future<Either<HttpException, MeEntity>> call(HttpParams params);
 }
