@@ -1,4 +1,4 @@
-import 'package:seller/src/modules/http/domain/helpers/http_helper.dart';
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
 
 import 'package:seller/src/core/domain/entities/exception/exception.dart';
 import 'package:seller/src/core/domain/entities/either/either_entity.dart';
@@ -6,24 +6,5 @@ import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 import 'package:seller/src/modules/auth/domain/entities/signup/signup_entity.dart';
 
 abstract class SignUpUseCase {
-  Future<Either<HttpException, SignUpEntity>> call(SignUpParams params);
-}
-
-class SignUpParams {
-  final String url;
-  final String email;
-  final String password;
-  final HttpMethod httpMethod;
-
-  const SignUpParams({
-    required this.url,
-    required this.email,
-    required this.password,
-    required this.httpMethod,
-  });
-
-  Map<String, dynamic> get toJson => {
-        'email': email,
-        'password': password,
-      };
+  Future<Either<HttpException, SignUpEntity>> call(HttpParams params);
 }
