@@ -1,4 +1,6 @@
-import 'package:provider/provider.dart';
+import 'package:get_it/get_it.dart';
+
+import 'package:seller/src/core/external/di/module.dart';
 
 import 'package:seller/src/modules/http/data/clients/http_client.dart';
 
@@ -10,34 +12,42 @@ import 'package:seller/src/modules/firabase/domain/usecases/crashlytics_usecase.
 import 'package:seller/src/modules/auth/data/usecases/me/remote_me_usecase.dart';
 import 'package:seller/src/modules/auth/data/usecases/signup/remote_signup_usecase.dart';
 
-import 'package:seller/src/modules/auth/presentation/screens/auth/me/states/me_logic.dart';
-import 'package:seller/src/modules/auth/presentation/screens/auth/signup/states/signup_logic.dart';
+import 'package:seller/src/modules/auth/presentation/screens/auth/me/bloc/me_bloc.dart';
+import 'package:seller/src/modules/auth/presentation/screens/auth/signup/bloc/signup_bloc.dart';
 
-final authModule = [
-  /// SignUp
-  Provider<SignUpUseCase>(
-    create: (context) => RemoteSignUpUseCase(
-      context.read<HttpClient>(),
-    ),
-  ),
-  Provider<SignUpLogic>(
-    create: (context) => SignUpLogic(
-      context.read<SignUpUseCase>(),
-      context.read<CrashlyticsUseCase>(),
-    ),
-  ),
+class AuthModule implements Module {
+  final GetIt locator;
 
-  /// Me
-  Provider<MeUseCase>(
-    create: (context) => RemoteMeUseCase(
-      context.read<HttpClient>(),
-    ),
-  ),
+  AuthModule(this.locator);
 
-  Provider<MeLogic>(
-    create: (context) => MeLogic(
-      context.read<MeUseCase>(),
-      context.read<CrashlyticsUseCase>(),
-    ),
-  ),
-];
+  @override
+  Future<void> initialized() async {
+    /// SignUp
+    locator
+      ..registerLazySingleton<SignUpUseCase>(
+        () => RemoteSignUpUseCase(
+          locator.get<HttpClient>(),
+        ),
+      )
+      ..registerFactory<SignupBloc>(
+        () => SignupBloc(
+          locator.get<SignUpUseCase>(),
+          locator.get<CrashlyticsUseCase>(),
+        ),
+      );
+
+    // Me
+    locator
+      ..registerLazySingleton<MeUseCase>(
+        () => RemoteMeUseCase(
+          locator.get<HttpClient>(),
+        ),
+      )
+      ..registerFactory<MeBloc>(
+        () => MeBloc(
+          locator.get<MeUseCase>(),
+          locator.get<CrashlyticsUseCase>(),
+        ),
+      );
+  }
+}
