@@ -1,66 +1,81 @@
 import 'package:flutter/material.dart';
 
-import 'package:provider/provider.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:seller/src/utils/constants/app_constant.dart';
 
-import 'package:seller/src/modules/http/domain/helpers/http_helper.dart';
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
 
-import 'package:seller/src/modules/auth/domain/usecases/signup/signup_usecase.dart';
+import 'package:seller/src/core/presentation/extensions/widget_extension.dart';
 
-import 'package:seller/src/modules/auth/presentation/screens/auth/me/me_screen.dart';
-import 'package:seller/src/modules/auth/presentation/screens/auth/signup/states/signup_logic.dart';
+import 'package:seller/src/modules/auth/presentation/routes/routers.dart';
+import 'package:seller/src/modules/auth/presentation/screens/auth/signup/bloc/signup_bloc.dart';
 
 class SignUpScreen extends StatelessWidget {
-  static const String path = '/sign_in';
-
-  final params = const SignUpParams(
-    password: 'cicada3301',
-    email: 'gjramos100@gmail.com',
+  final params = const HttpParams(
+    body: {
+      'email': 'gjramos100@gmail.com',
+      'password': 'cicada3301',
+    },
     url: AppContants.urlAuthSignUp,
-    httpMethod: HttpMethod.post,
   );
 
   const SignUpScreen({
     Key? key,
   }) : super(key: key);
 
+  void _navigateTo(String path, BuildContext context) {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      path,
+      (Route<dynamic> route) => false,
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    final _logic = context.read<SignUpLogic>();
+    final bloc = locator.get<SignupBloc>();
 
     return Scaffold(
-      body: ValueListenableBuilder(
-        valueListenable: _logic,
-        builder: (contex, state, __) {
-          if (state is SignUpLoadingState) {
-            return const Center(
-              child: CircularProgressIndicator.adaptive(),
-            );
+      body: BlocListener<SignupBloc, SignUpState>(
+        bloc: bloc,
+        listener: (_, SignUpState state) {
+          if (state is SignUpSuccessState) {
+            _navigateTo(AuthRoutes.pathMe, context);
           }
-
-          if (state is SignUpFailureState) {
-            return Center(
-              child: Text(state.message),
-            );
-          }
-
-          return Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                const Text('Sign UP'),
-                TextButton(
-                  onPressed: () async {
-                    await _logic.signUp(MeScreen.path, params, contex);
-                  },
-                  child: const Text('Enter'),
-                ),
-              ],
-            ),
-          );
         },
+        child: BlocBuilder<SignupBloc, SignUpState>(
+          bloc: bloc,
+          builder: (_, SignUpState state) {
+            if (state is SignUpLoadingState) {
+              return const Center(
+                child: CircularProgressIndicator.adaptive(),
+              );
+            }
+
+            if (state is SignUpFailureState) {
+              return Center(
+                child: Text(state.message),
+              );
+            }
+
+            return Center(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: <Widget>[
+                  const Text('Sign UP'),
+                  TextButton(
+                    onPressed: () {
+                      bloc.add(SignUpAwnerEvent(params));
+                    },
+                    child: const Text('Enter'),
+                  ),
+                ],
+              ),
+            );
+          },
+        ),
       ),
     );
   }
