@@ -1,29 +1,12 @@
 enum RemoteConfigResponse {
-  force,
-  intValue,
-  boolValue,
-  stringValue,
-  doubleValue,
-  dynamicValue
-}
+  force('helperNotHaveData'),
+  intValue('helperNotHaveData'),
+  boolValue('helperNotHaveData'),
+  stringValue('helperNotHaveData'),
+  doubleValue('helperNotHaveData'),
+  dynamicValue('helperNotHaveData');
 
-extension RemoteConfigResponseExtension on RemoteConfigResponse {
-  String get value {
-    switch (this) {
-      case RemoteConfigResponse.force:
-        return 'helperNotHaveData';
-      case RemoteConfigResponse.intValue:
-        return 'helperNotHaveData';
-      case RemoteConfigResponse.boolValue:
-        return 'helperNotHaveData';
-      case RemoteConfigResponse.stringValue:
-        return 'helperNotHaveData';
-      case RemoteConfigResponse.doubleValue:
-        return 'helperNotHaveData';
-      case RemoteConfigResponse.dynamicValue:
-        return 'helperNotHaveData';
-      default:
-        return 'helperGeneric';
-    }
-  }
+  final String value;
+
+  const RemoteConfigResponse(this.value);
 }
