@@ -1,5 +1,5 @@
-import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 import 'package:seller/src/core/domain/entities/exception/exception.dart';
+import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 
 abstract class RemoteConfigUseCase {
   Future<Either<RemoteConfigException, bool>> forceRefresh();
