@@ -1,15 +1,15 @@
-import 'package:seller/src/modules/http/domain/helpers/http_helper.dart';
-
+import 'package:seller/src/modules/http/domain/params/http_params.dart';
 import 'package:seller/src/core/domain/entities/exception/exception.dart';
 import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 
 abstract class HttpClient {
-  void registerInterceptors();
+  void configureHttp();
 
-  Future<Either<HttpException, Map<String, dynamic>>> call({
-    required String url,
-    required HttpMethod method,
-    Map<String, dynamic>? body,
-    Map<String, String>? headers,
+  Future<Either<HttpException, Map<String, dynamic>>> get({
+    required HttpParams params,
+  });
+
+  Future<Either<HttpException, Map<String, dynamic>>> post({
+    required HttpParams params,
   });
 }
