@@ -1,16 +1,25 @@
-import 'package:provider/provider.dart';
+import 'package:get_it/get_it.dart';
 
-import 'package:seller/src/modules/splash/presentation/states/splash_logic.dart';
+import 'package:seller/src/core/external/di/module.dart';
+
+import 'package:seller/src/modules/splash/presentation/bloc/splash_bloc.dart';
 
 import 'package:seller/src/modules/storage/domain/usecases/storage_usecase.dart';
 
 import 'package:seller/src/modules/connectivity/domain/usecases/connectivity_usecase.dart';
 
-final splashModule = [
-  Provider<SplashLogic>(
-    create: (context) => SplashLogic(
-      context.read<StorageUsecase>(),
-      context.read<ConnectivityUseCase>(),
-    ),
-  ),
-];
+class SplashModule implements Module {
+  final GetIt locator;
+
+  SplashModule(this.locator);
+
+  @override
+  Future<void> initialized() async {
+    locator.registerFactory<SplashBloc>(
+      () => SplashBloc(
+        locator.get<StorageUseCase>(),
+        locator.get<ConnectivityUseCase>(),
+      ),
+    );
+  }
+}
