@@ -1,7 +1,7 @@
 import 'package:seller/src/core/domain/entities/exception/exception.dart';
 import 'package:seller/src/core/domain/entities/either/either_entity.dart';
 
-abstract class StorageUsecase {
+abstract class StorageUseCase {
   Future<Either<StorageException, bool>> clear();
   Either<StorageException, String> get({required String key});
   Future<Either<StorageException, bool>> delete({required String key});
