@@ -10,11 +10,10 @@ abstract class AppContants {
       '$_keySharedPreferencesName@refresh_token';
 
   // Url's
+  static const String urlHowgarts = 'url_howgarts';
   static const String urlMe = '/auth/v1/user';
   static const String urlAuthSignUp = '/auth/v1/signup';
-  static const String urlHowgarts = 'https://naiwfnrkbtsmzllsvmxj.supabase.co';
 
   // Key Supabase
-  static const String apikeyAnon =
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im5haXdmbnJrYnRzbXpsbHN2bXhqIiwicm9sZSI6ImFub24iLCJpYXQiOjE2NTA2NDQ4MjMsImV4cCI6MTk2NjIyMDgyM30.nvmqdU2FSdTzKHJ5_AJX7MIQmgyeEs5xpzXvO0BVXcw';
+  static const String apikeyAnon = 'api_key_anon';
 }
